(Findbugs) Valid-until calculation fixed. Integer overflow.
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@14549 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/hoster/ShareHosterDe.java
@@ -157,7 +157,7 @@ public class ShareHosterDe extends PluginForHost {
             account.setValid(false);
             return ai;
         } else {
-            ai.setValidUntil(System.currentTimeMillis() + Integer.parseInt(expire) * 24 * 60 * 60 * 1001);
+            ai.setValidUntil(System.currentTimeMillis() + Long.parseLong(expire) * 24 * 60 * 60 * 1001);
         }
         ai.setStatus("Premium User");
         return ai;
